Huge refactor of the onboarding process
Each step is now represented using a class, and the wrapping class uses
a graph to know which step is next.
This allows us to define common behavior and re-use it/override it via
inheritance.
This also has the benefit of not chaining calls to next step at the end
of each step's method, avoiding gigantic function call stacks.