Onboarding flow changes (cf. full commit message)
Newly joined members are now greeted in the server channel before they
go through the process, not when they're done with it.
The greeting message now prompts them to look for a DM.
If the DM cannot be sent due to user's settings, fall back to a generic
message.
Do not greet members when onboarding is triggered via the command.